2016-07-22 v.01

+++ src/main/java/com/certiware/backend/model/common/ManpowerModel.java
@@ -4,15 +4,15 @@ import java.util.Date;
 
 public class ManpowerModel {
 	
-	private int projectId;
-	private String manpowerName;
-	private int partnerId;
-	private String ratingCode;
-	private int sellingAmount;
-	private int outsourcingAmount;
-	private Date startDate;
-	private Date endDate;
-	private String remarks;
+	private int projectId;			// 프로젝트 아이디
+	private String manpowerName;	// 투입인력이름
+	private int partnerId;			// 소속 아이디
+	private String ratingCode;		// 등급
+	private long sellingAmount;		// 매출단가
+	private long outsourcingAmount;	// 외주단가
+	private Date startDate;			// 시작일
+	private Date endDate;			// 종료일
+	private String remarks;			// 비고
 	public int getProjectId() {
 		return projectId;
 	}
@@ -37,16 +37,16 @@ public class ManpowerModel {
 	public void setRatingCode(String ratingCode) {
 		this.ratingCode = ratingCode;
 	}
-	public int getSellingAmount() {
+	public long getSellingAmount() {
 		return sellingAmount;
 	}
-	public void setSellingAmount(int sellingAmount) {
+	public void setSellingAmount(long sellingAmount) {
 		this.sellingAmount = sellingAmount;
 	}
-	public int getOutsourcingAmount() {
+	public long getOutsourcingAmount() {
 		return outsourcingAmount;
 	}
-	public void setOutsourcingAmount(int outsourcingAmount) {
+	public void setOutsourcingAmount(long outsourcingAmount) {
 		this.outsourcingAmount = outsourcingAmount;
 	}
 	public Date getStartDate() {
@@ -67,6 +67,7 @@ public class ManpowerModel {
 	public void setRemarks(String remarks) {
 		this.remarks = remarks;
 	}
+
